refactor: add payment provider resolver service and inject into customers
