+++ internal/services/src/tests/workflow.test.ts
@@ -8,6 +8,7 @@ import type { Cache } from "../cache/service"
 import { CustomerService } from "../customers/service"
 import { GrantsManager } from "../entitlements/grants"
 import type { Metrics } from "../metrics"
+import { PaymentProviderResolver } from "../payment-provider/resolver"
 import { SubscriptionService } from "../subscriptions/service"
 import { createClock } from "../test-utils"
 
@@ -281,7 +282,11 @@ describe("Workflow - Billing and Subscriptions", () => {
       metrics: mockMetrics,
     }
 
-    const customerService = new CustomerService(serviceDeps)
+    const paymentProviderResolver = new PaymentProviderResolver({
+      db: mockDb,
+      logger: mockLogger,
+    })
+    const customerService = new CustomerService({ ...serviceDeps, paymentProviderResolver })
     const grantsManager = new GrantsManager({ db: mockDb, logger: mockLogger })
     _billingService = new BillingService({ ...serviceDeps, customerService, grantsManager })
     subscriptionService = new SubscriptionService({
